Fixed duplication of images which are inserted at the end of a paragraph
