+++ richtextfx/src/main/java/org/fxmisc/richtext/StyledTextArea.java
@@ -66,8 +66,8 @@ public class StyledTextArea<PS, S> extends GenericStyledArea<PS, StyledText<S>, 
                           EditableStyledDocument<PS, StyledText<S>, S> document, boolean preserveStyle) {
         super(initialParagraphStyle, applyParagraphStyle,
               initialTextStyle,
-              document, StyledText.textOps(), preserveStyle,
-              seg -> createStyledTextNode(seg, StyledText.textOps(), applyStyle));
+              document, StyledText.textOps(initialTextStyle), preserveStyle,
+              seg -> createStyledTextNode(seg, StyledText.textOps(initialTextStyle), applyStyle));
     }
 
     public StyledTextArea(PS initialParagraphStyle, BiConsumer<TextFlow, PS> applyParagraphStyle,
