+++ richtextfx/src/main/java/org/fxmisc/richtext/model/ReadOnlyStyledDocument.java
@@ -78,17 +78,17 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
         m.reset();
         while(m.find()) {
             String s = str.substring(start, m.start());
-            res.add(new Paragraph<>(paragraphStyle, segmentOps, segmentOps.create(s, style)));
+            res.add(new NonEmptyParagraph<>(paragraphStyle, segmentOps, segmentOps.create(s, style)));
             start = m.end();
         }
         String last = str.substring(start);
-        res.add(new Paragraph<>(paragraphStyle, segmentOps, segmentOps.create(last, style)));
+        res.add(new NonEmptyParagraph<>(paragraphStyle, segmentOps, segmentOps.create(last, style)));
 
         return new ReadOnlyStyledDocument<>(res);
     }
 
     public static <PS, SEG, S> ReadOnlyStyledDocument<PS, SEG, S> fromSegment(SEG segment,  PS paragraphStyle, S style, SegmentOps<SEG, S> segmentOps) {
-        Paragraph<PS, SEG, S> content = new Paragraph<PS, SEG, S>(paragraphStyle, segmentOps, Arrays.asList(segment));
+        Paragraph<PS, SEG, S> content = new NonEmptyParagraph<>(paragraphStyle, segmentOps, Arrays.asList(segment));
         List<Paragraph<PS, SEG, S>> res = Arrays.asList(content);
         return new ReadOnlyStyledDocument<>(res);
     }
@@ -143,7 +143,7 @@ public final class ReadOnlyStyledDocument<PS, SEG, S> implements StyledDocument<
             public Paragraph<PS, SEG, S> decode(DataInputStream is) throws IOException {
                 PS paragraphStyle = pCodec.decode(is);
                 List<SEG> segments = segmentsCodec.decode(is);
-                return new Paragraph<>(paragraphStyle, segmentOps, segments);
+                return new NonEmptyParagraph<>(paragraphStyle, segmentOps, segments);
             }
         };
     }
