+++ richtextfx/src/main/java/org/fxmisc/richtext/model/SimpleEditableStyledDocument.java
@@ -7,6 +7,6 @@ package org.fxmisc.richtext.model;
 public final class SimpleEditableStyledDocument<PS, S> extends GenericEditableStyledDocumentBase<PS, StyledText<S>, S> {
 
     public SimpleEditableStyledDocument(PS initialParagraphStyle, S initialStyle) {
-        super(initialParagraphStyle, initialStyle, StyledText.textOps());
+        super(initialParagraphStyle, initialStyle, StyledText.textOps(initialStyle));
     }
 }
