+++ richtextfx/src/main/java/org/fxmisc/richtext/model/StyledText.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public class StyledText<S>  {
 
-    public static <S> TextOps<StyledText<S>, S> textOps() {
+    public static <S> TextOps<StyledText<S>, S> textOps(S defaultStyle) {
         return new TextOps<StyledText<S>, S>() {
 
             @Override
@@ -27,13 +27,18 @@ public class StyledText<S>  {
             }
 
             @Override
-            public StyledText<S> subSequence(StyledText<S> styledText, int start, int end) {
-                return new StyledText<>(styledText.getText().substring(start, end), styledText.getStyle());
+            public Optional<StyledText<S>> subSequence(StyledText<S> styledText, int start, int end) {
+                return Optional.of(new StyledText<>(styledText.getText().substring(start, end), styledText.getStyle()));
             }
 
             @Override
-            public StyledText<S> subSequence(StyledText<S> styledText, int start) {
-                return new StyledText<>(styledText.getText().substring(start), styledText.getStyle());
+            public Optional<StyledText<S>> subSequence(StyledText<S> styledText, int start) {
+                 return Optional.of(new StyledText<>(styledText.getText().substring(start), styledText.getStyle()));
+            }
+
+            @Override
+            public S defaultStyle() {
+                return defaultStyle;
             }
 
             @Override
