+++ richtextfx/src/main/java/org/fxmisc/richtext/model/TextOps.java
@@ -6,15 +6,19 @@ public interface TextOps<SEG, S> extends SegmentOps<SEG, S> {
     public SEG create(String text, S style);
 
     public default <R> TextOps<Either<SEG, R>, S> _or(SegmentOps<R, S> rOps) {
-        return eitherL(this, rOps);
+        return _or(rOps, EitherSegmentOps.StyleChoice.LEFT);
     }
 
-    public static <L, R, S> TextOps<Either<L, R>, S> eitherL(TextOps<L, S> lOps, SegmentOps<R, S> rOps) {
-        return new LeftTextOps<>(lOps, rOps);
+    public default <R> TextOps<Either<SEG, R>, S> _or(SegmentOps<R, S> rOps, EitherSegmentOps.StyleChoice choice) {
+        return eitherL(this, rOps, choice);
     }
 
-    public static <L, R, S> TextOps<Either<L, R>, S> eitherR(SegmentOps<L, S> lOps, TextOps<R, S> rOps) {
-        return new RightTextOps<>(lOps, rOps);
+    public static <L, R, S> TextOps<Either<L, R>, S> eitherL(TextOps<L, S> lOps, SegmentOps<R, S> rOps, EitherSegmentOps.StyleChoice choice) {
+        return new LeftTextOps<>(lOps, rOps, choice);
+    }
+
+    public static <L, R, S> TextOps<Either<L, R>, S> eitherR(SegmentOps<L, S> lOps, TextOps<R, S> rOps, EitherSegmentOps.StyleChoice choice) {
+        return new RightTextOps<>(lOps, rOps, choice);
     }
 }
 
@@ -22,8 +26,8 @@ class LeftTextOps<L, R, S> extends EitherSegmentOps<L, R, S> implements TextOps<
 
     private final TextOps<L, S> lOps;
 
-    LeftTextOps(TextOps<L, S> lOps, SegmentOps<R, S> rOps) {
-        super(lOps, rOps);
+    LeftTextOps(TextOps<L, S> lOps, SegmentOps<R, S> rOps, StyleChoice choice) {
+        super(lOps, rOps, choice);
         this.lOps = lOps;
     }
 
@@ -38,8 +42,8 @@ class RightTextOps<L, R, S> extends EitherSegmentOps<L, R, S> implements TextOps
 
     private final TextOps<R, S> rOps;
 
-    RightTextOps(SegmentOps<L, S> lOps, TextOps<R, S> rOps) {
-        super(lOps, rOps);
+    RightTextOps(SegmentOps<L, S> lOps, TextOps<R, S> rOps, StyleChoice choice) {
+        super(lOps, rOps, choice);
         this.rOps = rOps;
     }
