+++ richtextfx/src/test/java/org/fxmisc/richtext/model/ParagraphTest.java
@@ -1,6 +1,6 @@
 package org.fxmisc.richtext.model;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
 
@@ -11,9 +11,9 @@ public class ParagraphTest {
     // This relates to merging text changes and issue #216.
     @Test
     public void concatEmptyParagraphsTest() {
-        TextOps<StyledText<Boolean>, Boolean> segOps = StyledText.textOps();
-        Paragraph<Void, StyledText<Boolean>, Boolean> p1 = new Paragraph<>(null, segOps, segOps.create("", true));
-        Paragraph<Void, StyledText<Boolean>, Boolean> p2 = new Paragraph<>(null, segOps, segOps.create("", false));
+        TextOps<StyledText<Boolean>, Boolean> segOps = StyledText.textOps(false);
+        Paragraph<Void, StyledText<Boolean>, Boolean> p1 = new NonEmptyParagraph<>(null, segOps, segOps.create("", true));
+        Paragraph<Void, StyledText<Boolean>, Boolean> p2 = new NonEmptyParagraph<>(null, segOps, segOps.create("", false));
 
         Paragraph<Void, StyledText<Boolean>, Boolean> p = p1.concat(p2);
