+++ richtextfx/src/test/java/org/fxmisc/richtext/model/ReadOnlyStyledDocumentTest.java
@@ -13,7 +13,7 @@ public class ReadOnlyStyledDocumentTest {
 
     @Test
     public void testUndo() {
-        TextOps<StyledText<String>, String> segOps = StyledText.textOps();
+        TextOps<StyledText<String>, String> segOps = StyledText.textOps("");
         ReadOnlyStyledDocument<String, StyledText<String>, String> doc0 = fromString("", "X", "X", segOps);
 
         doc0.replace(0, 0, fromString("abcd", "Y", "Y", segOps)).exec((doc1, chng1, pchng1) -> {
@@ -30,16 +30,16 @@ public class ReadOnlyStyledDocumentTest {
 
     @Test
     public void deleteNewlineTest() {
-        TextOps<StyledText<Void>, Void> segOps = StyledText.textOps();
+        TextOps<StyledText<Void>, Void> segOps = StyledText.textOps(null);
         ReadOnlyStyledDocument<Void, StyledText<Void>, Void> doc0 = fromString("Foo\nBar", null, null, segOps);
         doc0.replace(3, 4, fromString("", null, null, segOps)).exec((doc1, ch, pch) -> {
             List<? extends Paragraph<Void, StyledText<Void>, Void>> removed = pch.getRemoved();
             List<? extends Paragraph<Void, StyledText<Void>, Void>> added = pch.getAdded();
             assertEquals(2, removed.size());
-            assertEquals(new Paragraph<Void, StyledText<Void>, Void>(null, segOps, segOps.create("Foo", null)), removed.get(0));
-            assertEquals(new Paragraph<Void, StyledText<Void>, Void>(null, segOps, segOps.create("Bar", null)), removed.get(1));
+            assertEquals(new NonEmptyParagraph<Void, StyledText<Void>, Void>(null, segOps, segOps.create("Foo", null)), removed.get(0));
+            assertEquals(new NonEmptyParagraph<Void, StyledText<Void>, Void>(null, segOps, segOps.create("Bar", null)), removed.get(1));
             assertEquals(1, added.size());
-            assertEquals(new Paragraph<Void, StyledText<Void>, Void>(null, segOps, segOps.create("FooBar", null)), added.get(0));
+            assertEquals(new NonEmptyParagraph<Void, StyledText<Void>, Void>(null, segOps, segOps.create("FooBar", null)), added.get(0));
         });
     }
 
@@ -48,7 +48,7 @@ public class ReadOnlyStyledDocumentTest {
         final String fooBar = "Foo Bar";
         final String and = " and ";
         final String helloWorld = "Hello World";
-        TextOps<StyledText<String>, String> segOps = StyledText.textOps();
+        TextOps<StyledText<String>, String> segOps = StyledText.textOps("");
 
         SimpleEditableStyledDocument<String, String> doc0 = new SimpleEditableStyledDocument<>("", "");
